Place token handling in command and session.
