+++ cypress.json
@@ -5,6 +5,7 @@
   "video": true,
   "videoUploadOnPasses": false,
   "experimentalFetchPolyfill": true,
+  "experimentalSessionSupport": true,
   "viewportWidth": 1820,
   "viewportHeight": 1280,
   "projectId": "snwitz",
