+++ cypress/integration/Editor/rich_block_text_editor_spec.js
@@ -6,14 +6,13 @@
  *
  */
 
-import { setToken } from '../../support';
 import editorRoutes from './editorRoutes';
 
 const ARTICLE_ID = 800;
 
 describe('Learning resource editing', () => {
   beforeEach(() => {
-    setToken();
+    cy.setToken();
     editorRoutes(ARTICLE_ID);
 
     cy.visit(`/subject-matter/learning-resource/${ARTICLE_ID}/edit/nb`);
