+++ cypress/integration/Search/search_content_spec.js
@@ -7,11 +7,10 @@
  */
 
 import { taxonomyApi } from '../../../src/config';
-import { setToken } from '../../support';
 
 describe('Search content', () => {
   beforeEach(() => {
-    setToken();
+    cy.setToken();
     cy.apiroute('GET', `${taxonomyApi}/resource-types/?language=nb`, 'resourceTypes');
     cy.apiroute('GET', `${taxonomyApi}/subjects?language=nb`, 'allSubjects');
     cy.apiroute('GET', '/search-api/v1/search/editorial/*', 'search');
