+++ package.json
@@ -66,7 +66,7 @@
     "chalk": "^2.4.2",
     "concurrently": "^5.3.0",
     "cross-env": "^5.2.0",
-    "cypress": "^7.2.0",
+    "cypress": "^8.2.0",
     "eslint": "^7.22.0",
     "eslint-config-ndla": "^2.0.3",
     "eslint-plugin-cypress": "^2.11.2",
